Rename the Python library to spreen-wiki
spreen — a blend of the falcon's stoop and the preen that follows, after
Hayato (隼人, falcon man). The PyPI distribution becomes spreen-wiki, the
import package spreen_wiki, the console script spreen and the config file
.spreen.yml.
